#88 Modify move refresh to axios
